pyproject.toml
[tool.poetry]
name = "prayers"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
redis = "^5.2.0"
uvicorn = "^0.32.1"
fastapi = "^0.115.5"
groq = "^0.12.0"
python-dotenv = "^1.0.1"
tweety-ns = "^1.1.9"
httpx = {extras = ["http2"], version = "^0.27.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
